requirements.txt
# ── Data ──────────────────────────────────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0

# ── LangChain + OpenAI-compatible client ──────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0      # provides ChatOpenAI with base_url support

# ── News search ───────────────────────────────────────────────────────────────
tavily-python>=0.3.0

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Display ───────────────────────────────────────────────────────────────────
colorama>=0.4.6
